Add find_removed_packages() function
Function find_removed_packages() scans the packages lists and extracts their
names only.

After extracted, it transforms them into sets for better performance and finds
the packages in the initial list and not in the final list. Those are the removed
packages.

Last, it prints the removed packages in the requested output format